requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
sentence-transformers
langchain-experimental
langchain-huggingface
langchain-community
langchain-core
google-generativeai
python-dotenv
pypdf
google
typing
fastapi